+++ lib/flutter_slides.dart
@@ -8,3 +8,4 @@ export 'package:flutter_slides/components/subtopic.dart';
 export 'package:flutter_slides/components/topic.dart';
 export 'package:flutter_slides/components/topic_box.dart';
 export 'package:flutter_slides/components/topic_space.dart';
+export 'package:flutter_slides/components/slide_title.dart';